+++ constants.py
@@ -1,0 +1,28 @@
+# Constants for Colors
+class Colors:
+    WHITE = "#FFFFFF"
+    BLACK = "#000000"
+    RED = "#FF0000"
+    GREEN = "#00FF00"
+    BLUE = "#0000FF"
+    CNFRMCOL = "#F6D6B1"
+    ORDCOL="#B47950"
+    ORDENTRY="#C4C4C4"
+    CONCOL="#f4fbfd"
+class Dimensions:
+#constants for Home page buttons
+#About us button
+    ABT_WIDTH = 219
+    ABT_HEIGHT = 62
+
+#ORDER NOW button
+    ORD_HEIGHT = 62
+    ORD_WIDTH = 267
+
+# Constants for Other Button Sizes
+    BUTTON_WIDTH = 153
+    BUTTON_HEIGHT = 53
+
+# Constants for Image Paths
+class Paths:
+    IMAGE_DIRECTORY = r"Images\\"
